+++ packages/snap-preact/src/Templates/Stores/TemplateStore.ts
@@ -12,7 +12,6 @@ import { TargetStore } from './TargetStore';
 import { CurrencyCodes, LanguageCodes, LibraryImports, LibraryStore } from './LibraryStore';
 import { debounce } from '@athoscommerce/snap-toolbox';
 import type { PluginFunction } from '@athoscommerce/snap-controller';
-import type { OverlayResultProps } from '../../../components/src/components/Molecules/OverlayResult';
 import type {
 	PluginAddToCartConfig as PluginShopifyAddToCartConfig,
 	PluginBackgroundFiltersConfig as PluginShopifyBackgroundFiltersConfig,
@@ -41,66 +40,7 @@ import type {
 	ThemeMinimal,
 	ThemeOverrides,
 	ThemeVariablesPartial,
-	// Component Props for typed component config
-	ResultProps,
-	OverlayBadgeProps,
-	BadgeImageProps,
-	BadgePillProps,
-	BadgeRectangleProps,
-	BadgeTextProps,
-	BreadcrumbsProps,
-	ButtonProps,
-	DropdownProps,
-	FormattedNumberProps,
-	IconProps,
-	ImageProps,
-	LoadingBarProps,
-	BannerProps,
-	InlineBannerProps,
-	OverlayProps,
-	PaginationInfoProps,
-	SlideshowProps,
-	PriceProps,
-	SkeletonProps,
-	ModalProps,
-	CalloutBadgeProps,
-	CarouselProps,
-	CheckboxProps,
-	GridProps,
-	LayoutSelectorProps,
-	ListProps,
-	RadioProps,
-	ErrorHandlerProps,
-	FacetGridOptionsProps,
-	FacetHierarchyOptionsProps,
-	FacetListOptionsProps,
-	FacetPaletteOptionsProps,
-	FacetSliderProps,
-	FilterProps,
-	LoadMoreProps,
-	PaginationProps,
-	PerPageProps,
-	RadioListProps,
-	RatingProps,
-	SearchInputProps,
-	SelectProps,
-	SlideoutProps,
-	SortByProps,
-	SwatchesProps,
-	VariantSelectionProps,
-	TermsProps,
-	BranchOverrideProps,
-	FacetProps,
-	FacetsProps,
-	FacetsHorizontalProps,
-	FilterSummaryProps,
-	NoResultsProps,
-	ResultsProps,
-	SearchHeaderProps,
-	SidebarProps,
-	MobileSidebarProps,
-	ToolbarProps,
-	TermsListProps,
+	ComponentTypePropsMap,
 } from '../../../components/src';
 import type { GlobalThemeStyleScript, IntegrationPlatforms } from '../../types';
 import type { ClientConfig } from '@athoscommerce/snap-client';
@@ -233,71 +173,6 @@ type TemplatesStoreThemeConfigLocked = {
 
 type TemplatesStoreThemeConfigUnlocked = Omit<TemplatesStoreThemeConfigLocked, 'overrides'> & {
 	overrides?: ThemeResponsiveCompleteUnlocked;
-};
-
-// Component type to props mapping for typed component config
-export type ComponentTypePropsMap = {
-	result: ResultProps;
-	badge: OverlayBadgeProps;
-	overlayResult: OverlayResultProps;
-	badgeImage: BadgeImageProps;
-	badgePill: BadgePillProps;
-	badgeRectangle: BadgeRectangleProps;
-	badgeText: BadgeTextProps;
-	breadcrumbs: BreadcrumbsProps;
-	button: ButtonProps;
-	dropdown: DropdownProps;
-	formattedNumber: FormattedNumberProps;
-	icon: IconProps;
-	image: ImageProps;
-	loadingBar: LoadingBarProps;
-	banner: BannerProps;
-	inlineBanner: InlineBannerProps;
-	overlay: OverlayProps;
-	paginationInfo: PaginationInfoProps;
-	slideshow: SlideshowProps;
-	price: PriceProps;
-	skeleton: SkeletonProps;
-	modal: ModalProps;
-	calloutBadge: CalloutBadgeProps;
-	carousel: CarouselProps;
-	checkbox: CheckboxProps;
-	grid: GridProps;
-	layoutSelector: LayoutSelectorProps;
-	list: ListProps;
-	radio: RadioProps;
-	errorHandler: ErrorHandlerProps;
-	facetGridOptions: FacetGridOptionsProps;
-	facetHierarchyOptions: FacetHierarchyOptionsProps;
-	facetListOptions: FacetListOptionsProps;
-	facetPaletteOptions: FacetPaletteOptionsProps;
-	facetSlider: FacetSliderProps;
-	filter: FilterProps;
-	loadMore: LoadMoreProps;
-	overlayBadge: OverlayBadgeProps;
-	pagination: PaginationProps;
-	perPage: PerPageProps;
-	radioList: RadioListProps;
-	rating: RatingProps;
-	searchInput: SearchInputProps;
-	select: SelectProps;
-	slideout: SlideoutProps;
-	sortBy: SortByProps;
-	swatches: SwatchesProps;
-	variantSelection: VariantSelectionProps;
-	terms: TermsProps;
-	branchOverride: BranchOverrideProps;
-	facet: FacetProps;
-	facets: FacetsProps;
-	facetsHorizontal: FacetsHorizontalProps;
-	filterSummary: FilterSummaryProps;
-	noResults: NoResultsProps;
-	results: ResultsProps;
-	searchHeader: SearchHeaderProps;
-	sidebar: SidebarProps;
-	mobileSidebar: MobileSidebarProps;
-	toolbar: ToolbarProps;
-	termsList: TermsListProps;
 };
 
 // Typed component function: returns a component that accepts the mapped props type
