+++ packages/snap-preact/src/Templates/index.ts
@@ -1,3 +1,3 @@
 export { SnapTemplates } from './SnapTemplates';
 export type { SnapTemplatesConfig, SnapTemplatesConfigUnlocked } from './SnapTemplates';
-export type { TemplatesStoreParams, ComponentTypePropsMap } from './Stores/TemplateStore';
+export type { TemplatesStoreParams } from './Stores/TemplateStore';
